#8 Store username on req.session object
