+++ app.js
@@ -5,8 +5,9 @@ const logger = require('morgan');
 const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
 const exphbs = require('express-handlebars');
-const router = require('./routes/index');
 const cookieSession = require('cookie-session');
+
+const router = require('./routes/index');
 
 const app = express();
 
@@ -24,15 +25,15 @@ app.engine('hbs', exphbs({
 // uncomment after placing your favicon in /public
 // app.use(favicon(path.join(__dirname, 'public', 'favicon.ico')));
 app.use(logger('dev'));
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(cookieParser());
-app.use(express.static(path.join(__dirname, 'public')));
 app.use(cookieSession({
   name: 'session',
   secret: process.env.SESSION_SECRET,
   maxAge: 24 * 60 * 60 * 1000, // 24 hours
 }));
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: false }));
+app.use(cookieParser());
+app.use(express.static(path.join(__dirname, 'public')));
 
 app.use(router);
