+++ routes/index.js
@@ -3,12 +3,15 @@ const router = express.Router();
 
 const report = require('./report');
 const scrapeLinks = require('./scrape-links');
-const links = require('./links');
+const githubAuth = require('./github-auth');
 
 const login = (req, res) => {
   res.render('login');
 };
 
+const links = (req, res) => {
+  res.render('scrape-form');
+};
 const validateForm = (req, res) => {
   res.render('validate-form');
 };
@@ -18,5 +21,7 @@ router.get('/links', links);
 router.get('/links-validate', validateForm);
 router.get('/scrape-links', scrapeLinks);
 router.get('/report', report);
+
+router.get('/auth', githubAuth);
 
 module.exports = router;
